DnD support for custom event components
